move scripts to scripts folder
