--- defaultServers.json
@@ -1,5 +1,0 @@
-{
-    "1": "wss://mainnet.infura.io/ws",
-    "3": "wss://ropsten.infura.io/ws",
-    "4": "wss://rinkeby.infura.io/ws"
-}
